requirements.txt
# django
Django==1.5.5

# django apps
django-floppyforms==1.1
South==0.7.6

# general libs
pillow==2.3.0
bleach==1.2.2

#search
Whoosh==2.6.0
django-haystack==2.1.0

# dev stuff
django-extensions==1.1.1
django-debug-toolbar==0.9.4
django-nose==1.1
django-taggit==0.11.2
ipython
--allow-external python-graph-core
--allow-unverified python-graph-core
python-graph-core==1.8.2
# fabric
# fabric-virtualenv
